+++ src/Commands/MakeTreeResourceCommand.php
@@ -56,7 +56,7 @@ class MakeTreeResourceCommand extends Command
 
     public static bool $shouldCheckModelsForSoftDeletes = true;
 
-    protected function configure()
+    protected function configure(): void
     {
         $this->addArgument(
             name: 'name',